Edits related to dbid / changeset management
change how we record the fact that we have processed an ibdf changeset
file (record in metadata store), move dbid down into the ochre folder,
start writing the dbid to the changeset folder and to a sememe on the
root concept.  

Add checking / failure if ids are out of sync.

Change the naming conventions of changeset files - create stable /
unique names

Enable rotation of changeset files after an arbitrary number of changes.

Allow setting the dbBuildMode / bootstrapMode from mojo