Lucene bug fixes / upgrades.  See changelog.
Updating Lucene version to current.  Upon start with an old database,
the indexes will be deleted, and new indexes will be built.  This will
take much longer than a normal startup.  The automatic reindex feature
is a new feature.


Bug fixed in lucene index design, where queries with sememe filters did
not find appropriate results on edits that had occurred after the
initial index.

Bugs fixed relating to timing issues on when an item was committed, vs
when it is available in the index.  Commit now does not complete until
the index is also up to date.